K8s API client moved into pkg/apiclient/k8s
