+++ pkg/apiclients/k8s/initclient.go
@@ -1,4 +1,4 @@
-package api
+package k8s
 
 import (
 	"fmt"
